+++ berserk/clients/puzzles.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import Iterator, Any, Dict, cast
+from typing import Iterator, Any, Dict, cast, List
 
 from .. import models
 from ..formats import NDJSON
@@ -78,3 +78,157 @@ class Puzzles(BaseClient):
         """
         path = "/api/racer"
         return cast(PuzzleRace, self._r.post(path))
+
+    def get_next(
+        self, theme: str = "mix", difficulty: str = "normal"
+    ) -> Dict[str, Any]:
+        """Get a random Lichess puzzle in JSON format. If authenticated,
+        only returns puzzles that the user has never seen before.
+
+         :param theme: The theme or opening to filter puzzles with.
+         :param difficulty: Enum:"easiest" "easier" "normal" "harder" "hardest"
+
+         :return: the next puzzle"""
+        path = "/api/puzzle/next"
+
+        allowed_themes: List[str] = [
+            "advancedPawn",
+            "advancedPawnDescription",
+            "advantage",
+            "advantageDescription",
+            "anastasiaMate",
+            "anastasiaMateDescription",
+            "arabianMate",
+            "arabianMateDescription",
+            "attackingF2F7",
+            "attackingF2F7Description",
+            "attraction",
+            "attractionDescription",
+            "backRankMate",
+            "backRankMateDescription",
+            "bishopEndgame",
+            "bishopEndgameDescription",
+            "bodenMate",
+            "bodenMateDescription",
+            "castling",
+            "castlingDescription",
+            "capturingDefender",
+            "capturingDefenderDescription",
+            "crushing",
+            "crushingDescription",
+            "doubleBishopMate",
+            "doubleBishopMateDescription",
+            "dovetailMate",
+            "dovetailMateDescription",
+            "equality",
+            "equalityDescription",
+            "kingsideAttack",
+            "kingsideAttackDescription",
+            "clearance",
+            "clearanceDescription",
+            "defensiveMove",
+            "defensiveMoveDescription",
+            "deflection",
+            "deflectionDescription",
+            "discoveredAttack",
+            "discoveredAttackDescription",
+            "doubleCheck",
+            "doubleCheckDescription",
+            "endgame",
+            "endgameDescription",
+            "enPassantDescription",
+            "exposedKing",
+            "exposedKingDescription",
+            "fork",
+            "forkDescription",
+            "hangingPiece",
+            "hangingPieceDescription",
+            "hookMate",
+            "hookMateDescription",
+            "interference",
+            "interferenceDescription",
+            "intermezzo",
+            "intermezzoDescription",
+            "knightEndgame",
+            "knightEndgameDescription",
+            "long",
+            "longDescription",
+            "master",
+            "masterDescription",
+            "masterVsMaster",
+            "masterVsMasterDescription",
+            "mate",
+            "mateDescription",
+            "mateIn1",
+            "mateIn1Description",
+            "mateIn2",
+            "mateIn2Description",
+            "mateIn3",
+            "mateIn3Description",
+            "mateIn4",
+            "mateIn4Description",
+            "mateIn5",
+            "mateIn5Description",
+            "middlegame",
+            "middlegameDescription",
+            "oneMove",
+            "oneMoveDescription",
+            "opening",
+            "openingDescription",
+            "pawnEndgame",
+            "pawnEndgameDescription",
+            "pin",
+            "pinDescription",
+            "promotion",
+            "promotionDescription",
+            "queenEndgame",
+            "queenEndgameDescription",
+            "queenRookEndgame",
+            "queenRookEndgameDescription",
+            "queensideAttack",
+            "queensideAttackDescription",
+            "quietMove",
+            "quietMoveDescription",
+            "rookEndgame",
+            "rookEndgameDescription",
+            "sacrifice",
+            "sacrificeDescription",
+            "short",
+            "shortDescription",
+            "skewer",
+            "skewerDescription",
+            "smotheredMate",
+            "smotheredMateDescription",
+            "superGM",
+            "superGMDescription",
+            "trappedPiece",
+            "trappedPieceDescription",
+            "underPromotion",
+            "underPromotionDescription",
+            "veryLong",
+            "veryLongDescription",
+            "xRayAttack",
+            "xRayAttackDescription",
+            "zugzwang",
+            "zugzwangDescription",
+            "mix",
+            "mixDescription",
+            "playerGames",
+            "playerGamesDescription",
+            "puzzleDownloadInformation",
+        ]
+
+        allowed_levels = ["easiest", "easy", "normal", "harder", "hardest"]
+
+        if theme not in allowed_themes:
+            raise KeyError("Entered theme not supported")
+
+        difficulty = difficulty.lower()
+        if difficulty not in allowed_levels:
+            raise KeyError(
+                "Please enter a difficulty which is one of "
+                "easiest, easier, normal, harder, hardest"
+            )
+
+        params = {"angle": theme, "difficulty": difficulty}
+        return self._r.get(path, params=params)